Commit on multi-disks
blockdev_commit_base: support commit on multi-disks
blockdev_commit: support commit on multi-disks
blockdev_commit: add cfg to support multi-disks
