+++ qemu/tests/cfg/blockdev_commit.cfg
@@ -1,10 +1,14 @@
 - blockdev_commit:
     type = blockdev_commit
     virt_test_type = qemu
+    only Linux
     images += " data"
-    force_create_image_data = yes
-    force_remove_image_data = yes
+    force_create_image = yes
+    force_remove_image = yes
+    force_create_image_image1 = no
+    force_remove_image_image1 = no
     start_vm = yes
+    kill_vm = yes
     storage_pools = default
     storage_type_default = "directory"
     storage_pool = default
@@ -30,10 +34,32 @@
 
     device_tag = "data"
     format = qcow2
-    #mode = "absolute-paths"
-    base_tag = sn1
-    top_tag = sn3
     rebase_mode = unsafe
     Host_RHEL.m8:
         node = ${device}
         qemu_force_use_drive_expression = no
+    variants:
+        - @one_data_disk:
+        - multi_data_disks:
+            image_size_data1 = 300M
+            image_name_data1 = data1
+            snapshot_tags_data1 = sn11 sn21 sn31 sn41
+
+            image_size_sn11 = 300M
+            image_name_sn11 = sn11
+            image_format_sn11 = qcow2
+
+            image_name_sn21 = sn21
+            image_size_sn21 = 300M
+            image_format_sn21 = qcow2
+
+            image_name_sn31 = sn31
+            image_size_sn31 = 300M
+            image_format_sn31 = qcow2
+
+            image_name_sn41 = sn41
+            image_size_sn41 = 300M
+            image_format_sn41 = qcow2
+
+            device_tag += " data1"
+            images += " data1"
